Split network/dcerpc by DCE/RPC protocol version (v4/v5)
Reorganize network/dcerpc to reflect the two DCE/RPC wire protocols defined
in C706: connection-oriented (rpc_vers = 5) and connectionless/datagram
(rpc_vers = 4).

- Move the entire existing connection-oriented implementation under
  network/dcerpc/v5/ (client, pdu, transport, transport/smb, interfaces).
  The package name stays "dcerpc" per Go's /vN convention, so exported
  identifiers (dcerpc.MajorVersion, ...) are unchanged; only import-path
  strings change. History preserved via git mv.
- Hoist the version-agnostic packages up to shared locations,
  network/dcerpc/ndr and network/dcerpc/syntax, since the NDR transfer
  syntax and syntax identifiers are common to both protocol versions and
  must be importable by v4 without depending on v5.
- Add network/dcerpc/v4/ as a placeholder package (dcerpccl) documenting the
  connectionless protocol: the single rpc_vers field (4 LSB, no minor
  version), the 80-octet fixed header, and the 11 connectionless PDU types
  with their ptype values, cross-checked against C706 chapter 12. No
  connectionless encoding, transport, or client logic is implemented yet.
- Add C706 (chapters 12 and 10) and MS-RPCE reference URLs to the v4 and v5
  package doc comments.

No behavioral change: go build ./..., go vet, and the full
network/dcerpc/... test suite pass unchanged.
