+++ network/dcerpc/v5/interfaces/srvsvc/srvsvc.go
@@ -20,7 +20,7 @@ package srvsvc
 import (
 	"fmt"
 
-	"github.com/TheManticoreProject/Manticore/network/dcerpc/client"
+	"github.com/TheManticoreProject/Manticore/network/dcerpc/v5/client"
 	"github.com/TheManticoreProject/Manticore/network/dcerpc/ndr"
 	"github.com/TheManticoreProject/Manticore/network/dcerpc/syntax"
 	"github.com/TheManticoreProject/Manticore/windows/guid"
