+++ network/dcerpc/v5/transport/smb/smb_namedpipe_test.go
